PRO-4681: Refactoring
