--- src/test/end-to-end/pages/caseDetails/caseDetailsConfig.json
@@ -1,3 +1,0 @@
-{
-"goButton": "form button[type=\"submit\"]"
-}
