+++ src/test/end-to-end/pages/caseDetails/willLodgement/caseMatchesTabConfig.json
@@ -1,6 +1,5 @@
 {
   "tabName": "Case matches",
-  "group": "6",
   "fields": [
     "Full name",
     "Date of birth",
