+++ src/test/end-to-end/pages/caseDetails/willLodgement/documentUploadTabConfig.json
@@ -1,6 +1,5 @@
 {
   "tabName": "Documents",
-  "group": "2",
   "fields": [
     "Type",
     "Document Url",
