+++ src/test/end-to-end/pages/caseDetails/willLodgement/generateDepositReceiptTabConfig.json
@@ -1,6 +1,5 @@
 {
   "tabName": "History",
-  "group": "5",
   "fields": [
     "Date",
     "Author",
