+++ src/test/end-to-end/pages/caseMatches/caseMatchesConfig.json
@@ -1,9 +1,7 @@
 {
   "waitForText" : "Match application",
-  "pageUrl": "/create/case",
   "fullname": "testator_fn_update testator_ln_update",
   "dob": "1935-07-12",
   "dod": "2017-01-01",
-  "postcode": "SW1A 1AA",
-  "continueButton": "div.form-group.form-group-related > button[type=\"submit\"]"
+  "postcode": "SW1A 1AA"
 }
