+++ src/test/end-to-end/pages/checkYourAnswers/checkYourAnswers.js
@@ -2,6 +2,7 @@
 
 const testConfig = require('src/test/config.js');
 const checkYourAnswersConfig = require('./checkYourAnswersConfig.json');
+const commonConfig = require('src/test/end-to-end/pages/common/commonConfig');
 
 module.exports = function () {
 
@@ -11,5 +12,5 @@ module.exports = function () {
     I.fillField('#field-trigger-summary', checkYourAnswersConfig.summary);
     I.fillField('#field-trigger-description', checkYourAnswersConfig.comment);
 
-    I.waitForNavigationToComplete(checkYourAnswersConfig.locator);
+    I.waitForNavigationToComplete(commonConfig.continueButton);
 };
