+++ src/test/end-to-end/pages/common/commonConfig.json
@@ -1,0 +1,4 @@
+{
+  "goButton": "form button[type=\"submit\"]",
+  "continueButton": "div.form-group.form-group-related > button[type=\"submit\"]"
+}
