+++ src/test/end-to-end/pages/createCase/createCase.js
@@ -7,12 +7,11 @@ module.exports = function (jurisdiction, caseType, event) {
 
     const I = this;
     I.waitForText(createCaseConfig.waitForText, testConfig.TestTimeToWaitForText);
-    //  I.amOnPage(createCaseConfig.pageUrl)
 
     I.wait(10);
     I.selectOption('#cc-jurisdiction', jurisdiction);
     I.selectOption('#cc-case-type', caseType);
     I.selectOption('#cc-event', event);
 
-    I.waitForNavigationToComplete(createCaseConfig.locator);
+    I.waitForNavigationToComplete(createCaseConfig.startButton);
 };
