+++ src/test/end-to-end/pages/createWillLodgement/page1.js
@@ -2,6 +2,7 @@
 
 const testConfig = require('src/test/config');
 const createWillLodgementConfig = require('./createWillLodgementConfig');
+const commonConfig = require('src/test/end-to-end/pages/common/commonConfig');
 
 module.exports = function (crud) {
 
@@ -39,5 +40,5 @@ module.exports = function (crud) {
 
     }
 
-    I.click(createWillLodgementConfig.continueButton);
+    I.click(commonConfig.continueButton);
 };
