+++ src/test/end-to-end/pages/documentUpload/documentUploadConfig.json
@@ -1,10 +1,8 @@
 {
   "waitForText" : "Document upload",
-  "pageUrl": "/create/case",
   "addNewButton": " #documentsUploaded > div > button",
   "documentType": "Email",
   "fileToUploadUrl":"uploadDocuments/test_file_for_document_upload.png",
   "filename": "test_file_for_document_upload.png",
-  "comment": "test file to upload",
-  "continueButton": "div.form-group.form-group-related > button[type=\"submit\"]"
+  "comment": "test file to upload"
 }
