--- src/test/end-to-end/pages/eventSummary/caseMatchesCommentSummaryConfig.json
@@ -1,7 +1,0 @@
-{
-  "waitForText" : "Match application",
-  "pageUrl": "/create/case",
-  "summary": "case_match_application_event_summary",
-  "comment": "case_match_application_event_description_text",
-  "locator": "div.form-group.form-group-related > button[type=\"submit\"]"
-}
