--- src/test/end-to-end/pages/eventSummary/documentUploadSummaryConfig.json
@@ -1,7 +1,0 @@
-{
-  "waitForText" : "Upload document",
-  "pageUrl": "/create/case",
-  "summary": "upload_document_event_summary_purpose_text",
-  "comment": "upload_document_event_description_text",
-  "locator": "div.form-group.form-group-related > button[type=\"submit\"]"
-}
