+++ src/test/end-to-end/pages/eventSummary/eventSummary.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const testConfig = require('src/test/config.js');
+const commonConfig = require('src/test/end-to-end/pages/common/commonConfig');
 
 module.exports = function (caseRef, configFile) {
 
@@ -12,6 +13,6 @@ module.exports = function (caseRef, configFile) {
     I.fillField('#field-trigger-summary', configFile.summary);
     I.fillField('#field-trigger-description', configFile.comment);
 
-    I.waitForNavigationToComplete(configFile.locator);
+    I.waitForNavigationToComplete(commonConfig.continueButton);
 
 };
