+++ src/test/end-to-end/pages/eventSummary/willLodgement/caseMatchesCommentSummaryConfig.json
@@ -1,0 +1,5 @@
+{
+  "waitForText" : "Match application",
+  "summary": "case_match_application_event_summary",
+  "comment": "case_match_application_event_description_text"
+}
