+++ src/test/end-to-end/pages/eventSummary/willLodgement/commentSummaryConfig.json
@@ -1,0 +1,5 @@
+{
+  "waitForText" : "Add comment",
+  "summary": "comment_event_summary_purpose_text",
+  "comment": "comment_event_description_text"
+}
