+++ src/test/end-to-end/pages/eventSummary/willLodgement/generateDepositReceiptSummaryConfig.json
@@ -1,0 +1,5 @@
+{
+  "waitForText" : "Generate deposit receipt",
+  "summary": "deposit_receipt_event_summary_purpose_text",
+  "comment": "deposit_receipt_event_description_text"
+}
