--- src/test/end-to-end/pages/eventSummary/withdrawalSummaryConfig.json
@@ -1,7 +1,0 @@
-{
-  "waitForText" : "Withdraw will",
-  "pageUrl": "/create/case",
-  "summary": "withdrawal_event_summary_purpose_text",
-  "comment": "withdrawal_event_description_text",
-  "locator": "div.form-group.form-group-related > button[type=\"submit\"]"
-}
