+++ src/test/end-to-end/pages/newCase/newCase.js
@@ -2,14 +2,11 @@
 
 const testConfig = require('src/test/config');
 const newCaseConfig = require('./newCaseConfig.json');
-const {forEach} = require('lodash');
 
 module.exports = function () {
 
-        const I = this;
+    const I = this;
 
-        I.waitForText(newCaseConfig.waitForText,testConfig.TestTimeToWaitForText);
-      //  I.amOnPage(newCaseConfig.pageUrl);
-
-        I.waitForNavigationToComplete(newCaseConfig.locator);
+    I.waitForText(newCaseConfig.waitForText, testConfig.TestTimeToWaitForText);
+    I.waitForNavigationToComplete(newCaseConfig.locator);
 };
