+++ src/test/end-to-end/pages/withdrawal/withdrawalConfig.json
@@ -1,6 +1,4 @@
 {
   "waitForText" : "Will withdrawl",
-  "pageUrl": "/create/case",
-  "list1_text": "Cancelled",
-  "continueButton": "div.form-group.form-group-related > button[type=\"submit\"]"
+  "list1_text": "Cancelled"
 }
